Move LockFileLock to separate file
